+++ src/components/commands/php.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 ensure_php_is_installed() {
-    installApplicationHomebrewStyle "php@7.4"
-    brew link --force --overwrite php@7.4
+    installApplicationHomebrewStyle "php@8.0"
+    brew link --force --overwrite php@8.0
     brew services restart php
     brew unlink php && brew link php
 }