+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 ### Added
 - Install rbenv
+- Automatically add default SSH key, generating if required
 
 ## [1.5.0] - 2020-11-11
 ### Added